Readded Species death, balanced paper production
Turned PaperValue int into float

+++ Assets/Scripts/UI Scripts/UI_Shop.cs
@@ -96,7 +96,7 @@ public class UI_Shop : MonoBehaviour
         foreach (SpeciesType p in Resources.FindObjectsOfTypeAll(typeof(SpeciesType)) as SpeciesType[])
         {
             
-            CreateItemButton(p, p.SpeciesCost[0].PaperAmount, count, p.SpeciesCost[0].PaperColor);
+            CreateItemButton(p, (int) p.SpeciesCost[0].PaperAmount, count, p.SpeciesCost[0].PaperColor);
          
             
             count++;
@@ -151,7 +151,7 @@ public class UI_Shop : MonoBehaviour
     }
 
     public void ShopClick(SpeciesType s){
-        int paperNeeded = s.SpeciesCost[0].PaperAmount;
+        int paperNeeded = (int) s.SpeciesCost[0].PaperAmount;
         int paperHad = 0;
         
         if(s.SpeciesCost[0].PaperColor.PaperName == "blue")
